requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
pyyaml
psycopg2-binary
apache-airflow